pyproject.toml
[tool.poetry]
name = "ze-code-challenges"
version = "0.1.0"
description = "Backend challenge"
authors = ["Matheus Abdias <matheusabdias@b2bit.company>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
ipython = "^8.4.0"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
django-model-utils = "^4.2.0"
djangorestframework = "^3.13.1"
djangorestframework-gis = "^1.0"
django-filter = "^22.1"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
model-bakery = "^1.7.0"

[tool.poetry.dev-dependencies]
pylint = "^2.15.0"
flake8 = "^5.0.4"
black = {version = "^22.6.0", allow-prereleases = true}

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ze_delivery_challenge.settings"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
